Separate concerns for database
